i think i made things better
panning is now a thing and i think the code is more organized. idk im
tired it might be shit

+++ include/components/body.hpp
@@ -3,8 +3,8 @@
 #include <string>
 
 struct Body {
-  double mass;
-  double radius;
+  long double mass;
+  long double radius;
   std::string name;
 
   Body(double mass, double radius, std::string name)
