+++ include/components/camera.hpp
@@ -3,7 +3,7 @@
 #define ZOOM_DEFAULT 1.0 / 3e8
 
 struct Camera {
-    int x, y;
+    long long int x, y;
     double zoom;
 
     Camera() : x(0), y(0), zoom(ZOOM_DEFAULT) {}
